+++ src/main/java/kitchenpos/order/domain/Order.java
@@ -1,5 +1,7 @@
 package kitchenpos.order.domain;
 
+import kitchenpos.ordertable.domain.OrderTable;
+
 import javax.persistence.*;
 import java.time.LocalDateTime;
 import java.util.List;
